added all song pages
