+++ src/app/intania_news/intania_playlist/page.tsx
@@ -34,9 +34,9 @@ export default function IntaniaPlaylist() {
           <h1 className="text-center text-[22px] font-bold text-blue-1">1</h1>
         </div>
         <h1 className="text-center text-[18px] font-bold text-blue-1">
-          ปราสาทสีแดง
+          ปราสาทแดง
         </h1>
-        <Link href="/intania_news/intania_playlist/anthem">
+        <Link href="/intania_news/intania_playlist/song01">
           <button
             className="inline-flex items-center justify-center rounded-full bg-white p-2"
             style={{
@@ -58,41 +58,93 @@ export default function IntaniaPlaylist() {
           <h1 className="text-center text-[22px] font-bold text-blue-1">2</h1>
         </div>
         <h1 className="text-center text-[18px] font-bold text-blue-1">
-          ปราสาทสีแดง
+          แรงเลือดหมู
         </h1>
-        <button
-          className="inline-flex items-center justify-center rounded-full bg-white p-2"
-          style={{
-            height: "30px",
-            width: "30px",
-            borderRadius: "50px",
-            color: "red",
-          }}
-        >
-          <FaPlay style={{ fontSize: "11px" }} />
-        </button>
+        <Link href="/intania_news/intania_playlist/song02">
+          <button
+            className="inline-flex items-center justify-center rounded-full bg-white p-2"
+            style={{
+              height: "30px",
+              width: "30px",
+              borderRadius: "50px",
+              color: "red",
+            }}
+          >
+            <FaPlay style={{ fontSize: "11px" }} />
+          </button>
+        </Link>
       </div>
 
       <hr className="mb-3 mt-3 border-t-2 border-gray-600" />
 
-      <div className="mb-80 flex items-center space-x-4">
+      <div className="flex items-center space-x-4">
         <div className="w-10">
           <h1 className="text-center text-[22px] font-bold text-blue-1">3</h1>
         </div>
         <h1 className="text-center text-[18px] font-bold text-blue-1">
-          ปราสาทสีแดง
+          มหาจุฬาลงกรณ์
         </h1>
-        <button
-          className="inline-flex items-center justify-center rounded-full bg-white p-2"
-          style={{
-            height: "30px",
-            width: "30px",
-            borderRadius: "50px",
-            color: "red",
-          }}
-        >
-          <FaPlay style={{ fontSize: "11px" }} />
-        </button>
+        <Link href="/intania_news/intania_playlist/song03">
+          <button
+            className="inline-flex items-center justify-center rounded-full bg-white p-2"
+            style={{
+              height: "30px",
+              width: "30px",
+              borderRadius: "50px",
+              color: "red",
+            }}
+          >
+            <FaPlay style={{ fontSize: "11px" }} />
+          </button>
+        </Link>
+      </div>
+
+      <hr className="mb-3 mt-3 border-t-2 border-gray-600" />
+
+      <div className="flex items-center space-x-4">
+        <div className="w-10">
+          <h1 className="text-center text-[22px] font-bold text-blue-1">4</h1>
+        </div>
+        <h1 className="text-center text-[18px] font-bold text-blue-1">
+          บูมวิศวฯ (บูมบากา)
+        </h1>
+        <Link href="/intania_news/intania_playlist/song04">
+          <button
+            className="inline-flex items-center justify-center rounded-full bg-white p-2"
+            style={{
+              height: "30px",
+              width: "30px",
+              borderRadius: "50px",
+              color: "red",
+            }}
+          >
+            <FaPlay style={{ fontSize: "11px" }} />
+          </button>
+        </Link>
+      </div>
+
+      <hr className="mb-3 mt-3 border-t-2 border-gray-600" />
+
+      <div className="mb-48 flex items-center space-x-4">
+        <div className="w-10">
+          <h1 className="text-center text-[22px] font-bold text-blue-1">5</h1>
+        </div>
+        <h1 className="text-center text-[18px] font-bold text-blue-1">
+          วิศวกรรมเริงรื่น
+        </h1>
+        <Link href="/intania_news/intania_playlist/song05">
+          <button
+            className="inline-flex items-center justify-center rounded-full bg-white p-2"
+            style={{
+              height: "30px",
+              width: "30px",
+              borderRadius: "50px",
+              color: "red",
+            }}
+          >
+            <FaPlay style={{ fontSize: "11px" }} />
+          </button>
+        </Link>
       </div>
     </div>
   );
